Added check for InCopy and InDesign

+++ Work Scripts (Jamf Pro)/Uninstall Applications/Adobe/Adobe CC Apps/UninstallAllPreviousVersions_AdobeCCApp.sh
@@ -120,7 +120,9 @@ uninstallResult2017="$?"
 if [[ "$uninstallResult2017" -eq "0" ]]; then
     echo "${appNameForRemoval} 2017 uninstalled"
 fi
-rm -rf "/Applications/${appNameForRemoval} 2017" >/dev/null 2>&1
+if [[ -d "/Applications/${appNameForRemoval} 2017" ]]; then
+    rm -rf "/Applications/${appNameForRemoval} 2017" >/dev/null 2>&1
+fi
 # Uninstall 2018
 "$binaryPath" --uninstall=1 --sapCode="$sapCode" --baseVersion="$version2018" --platform=osx10-64 --deleteUserPreferences=false >/dev/null 2>&1
 uninstallResult2018="$?"
@@ -129,12 +131,16 @@ if [[ "$uninstallResult2018" -eq "0" ]]; then
 fi
 rm -rf "/Applications/${appNameForRemoval} 2018" >/dev/null 2>&1
 # Uninstall 2019
-"$binaryPath" --uninstall=1 --sapCode="$sapCode" --baseVersion="$version2019" --platform=osx10-64 --deleteUserPreferences=false >/dev/null 2>&1
-uninstallResult2019="$?"
-if [[ "$uninstallResult2019" -eq "0" ]]; then
-    echo "${appNameForRemoval} 2019 uninstalled"
+if [[ "$appNameForRemoval" =~ "InCopy" ]] || [[ "$appNameForRemoval" =~ "InDesign" ]]; then
+    echo "Policy is for ${appNameForRemoval} so 2019 version will not be removed"
+else
+    "$binaryPath" --uninstall=1 --sapCode="$sapCode" --baseVersion="$version2019" --platform=osx10-64 --deleteUserPreferences=false >/dev/null 2>&1
+    uninstallResult2019="$?"
+    if [[ "$uninstallResult2019" -eq "0" ]]; then
+        echo "${appNameForRemoval} 2019 uninstalled"
+    fi
+    rm -rf "/Applications/${appNameForRemoval} 2019" >/dev/null 2>&1
 fi
-rm -rf "/Applications/${appNameForRemoval} 2019" >/dev/null 2>&1
 }
 
 ########################################################################
